Add extension runtime and fix overlay interaction styling
