+++ packages/mesurer/index.ts
@@ -1,1 +1,10 @@
 export { default as Measurer } from "./measurer";
+export { mountMeasurer } from "./mount";
+export type {
+  MeasurerController,
+  MeasurerPortalTarget,
+  MeasurerProps,
+  MeasurerStorage,
+  MeasurerStyleTarget,
+  MountMeasurerOptions,
+} from "./api";
