+++ packages/mesurer/style-inject.ts
@@ -1,39 +1,55 @@
+import type { MeasurerStyleTarget } from "./api";
+
 const MESURER_STYLE_ID = "mesurer-styles";
 
-type StyleTarget = Document | ShadowRoot;
+const styleRefCounts = new WeakMap<
+  Node,
+  { count: number; node: HTMLStyleElement }
+>();
 
-const getStyleTarget = (
-  target?: HTMLElement | ShadowRoot,
-): StyleTarget | null => {
+const resolveStyleHost = (target?: MeasurerStyleTarget | null) => {
   if (typeof document === "undefined") return null;
-  if (!target) return document;
-  if (target instanceof ShadowRoot) return target;
-
-  const rootNode = target.getRootNode();
-  if (rootNode instanceof ShadowRoot) return rootNode;
-
-  return document;
+  if (!target) return document.head ?? document.documentElement;
+  if (target instanceof Document) return target.head ?? target.documentElement;
+  return target;
 };
 
 export function ensureMeasurerStyles(
   cssText: string,
-  target?: HTMLElement | ShadowRoot,
+  target?: MeasurerStyleTarget | null,
 ) {
-  if (typeof document === "undefined") return;
-  if (!cssText) return;
+  const host = resolveStyleHost(target);
+  if (!host || !cssText) return () => {};
 
-  const styleTarget = getStyleTarget(target);
-  if (!styleTarget) return;
-  if (styleTarget.querySelector(`#${MESURER_STYLE_ID}`)) return;
+  let entry = styleRefCounts.get(host);
+  if (!entry) {
+    const existing = host.querySelector(`#${MESURER_STYLE_ID}`);
+    const style =
+      existing instanceof HTMLStyleElement
+        ? existing
+        : document.createElement("style");
 
-  const style = document.createElement("style");
-  style.id = MESURER_STYLE_ID;
-  style.textContent = cssText;
+    style.id = MESURER_STYLE_ID;
+    style.textContent = cssText;
 
-  if (styleTarget instanceof ShadowRoot) {
-    styleTarget.appendChild(style);
-    return;
+    if (!style.parentNode) {
+      host.appendChild(style);
+    }
+
+    entry = { count: 0, node: style };
+    styleRefCounts.set(host, entry);
   }
 
-  styleTarget.head.appendChild(style);
+  entry.count += 1;
+
+  return () => {
+    const current = styleRefCounts.get(host);
+    if (!current) return;
+
+    current.count -= 1;
+    if (current.count > 0) return;
+
+    current.node.remove();
+    styleRefCounts.delete(host);
+  };
 }
