+++ packages/mesurer/styles.css
@@ -10,7 +10,8 @@
 }
 
 :root,
-:host {
+:host,
+.mesurer-root {
   --color-ink-50: #f8fafc;
   --color-ink-200: #e2e8f0;
   --color-ink-500: #64748b;
@@ -30,5 +31,34 @@
 }
 
 .mesurer-root {
+  color: var(--color-ink-900);
+  font-family:
+    Inter,
+    ui-sans-serif,
+    system-ui,
+    -apple-system,
+    BlinkMacSystemFont,
+    "Segoe UI",
+    sans-serif;
+  font-size: 16px;
+  line-height: 1.2;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
   --msr-spacing: 4px;
+}
+
+.mesurer-root,
+.mesurer-root *,
+.mesurer-root *::before,
+.mesurer-root *::after {
+  box-sizing: border-box;
+}
+
+.mesurer-root button,
+.mesurer-root input,
+.mesurer-root select,
+.mesurer-root textarea,
+.mesurer-root kbd {
+  font: inherit;
 }
